Split statistics in diferent functions
