+++ src/components/stats/GraphByCategory.tsx
@@ -5,7 +5,7 @@ import { round } from "../../domain/utils/round";
 import { useCategories } from "../../hooks/useCategories";
 import { Categories } from "../../interfaces";
 import CircularProgress from "@material-ui/core/CircularProgress";
-import { BudgetStatsByCategory } from "../../domain/BudgetStats";
+import { getTotalsByCategory } from "../../domain/stats/getTotalsByCategory";
 
 interface GraphByCategoryProps {
     budget: BudgetModel, 
@@ -23,7 +23,7 @@ export const GraphByCategory: React.FC<GraphByCategoryProps> = (props) => {
 
     const data = React.useMemo(() => {
         if (categoriesMap && Object.keys(categoriesMap).length > 0) {
-            const totals = new BudgetStatsByCategory(budget).totalsByCategory;
+            const totals = getTotalsByCategory(budget);
             const indexes = totals.indexes;
             const ignoreThreshold = totals.total * 0.03;
             return indexes
