+++ src/components/stats/GraphByCountry.tsx
@@ -1,8 +1,8 @@
 import * as React from "react";
 import { GraphPie } from "./Graph";
 import { round } from "../../domain/utils/round";
-import { BudgetStatsByCountry } from "../../domain/BudgetStats";
 import { BudgetModel } from "../../domain/BudgetModel";
+import { getTotalsByCountry } from "../../domain/stats/getTotalsByCountry";
 
 interface GraphByCountryProps {
     budget: BudgetModel
@@ -12,7 +12,7 @@ export const GraphByCountry: React.FC<GraphByCountryProps> = (props) => {
     const {budget} = props;
 
     const data = React.useMemo(() => {
-        const totals = new BudgetStatsByCountry(budget).totalsByCountry;
+        const totals = getTotalsByCountry(budget);
         const indexes = totals.indexes;
         const ignoreThreshold = totals.total * 0.05;
         return indexes
